Fix some warnings

+++ rpcs3/Emu/Cell/lv2/sys_ss.cpp
@@ -553,7 +553,7 @@ error_code sys_ss_individual_info_manager(u64 pkg_id, u64 a2, vm::ptr<u64> out_s
 	case 0x17002:
 	{
 		// TODO
-		vm::write<u64>(a5, a4); // Write back size of buffer
+		vm::write<u64>(static_cast<u32>(a5), a4); // Write back size of buffer
 		break;
 	}
 	// Get EID size
